requirements.txt
mysqlclient==2.0.3
Django==3.2
djangorestframework==3.12.4